project/requirements.txt
asgiref==3.3.0
Django==3.0.10
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
